Extracted the event repo tests out into an abstract test.
